requirements.txt
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
bcrypt==3.2.2
beanie==1.11.6
cffi==1.15.0
click==8.0.3
cryptography==37.0.2
Deprecated==1.2.13
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.73.0
fastapi-users==10.0.7
fastapi-users-db-beanie==1.1.0
h11==0.13.0
idna==3.3
makefun==1.13.1
motor==2.5.1
multidict==6.0.2
packaging==21.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.4.0
pymongo==3.12.3
pyparsing==3.0.9
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
redis==4.3.3
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.1.1
uvicorn==0.17.4
wrapt==1.14.1
yarl==1.7.2
